move placeholder type checking logic to PipelineTask to prevent v1 load breaks
