+++ sdk/python/kfp/components/pipeline_task.py
@@ -14,8 +14,9 @@
 """Pipeline task class and operations."""
 
 import copy
+import itertools
 import re
-from typing import Any, List, Mapping, Optional, Union
+from typing import Any, Dict, List, Mapping, Optional, Union
 
 from kfp.components import constants
 from kfp.components import pipeline_channel
@@ -128,7 +129,18 @@ class PipelineTask:
         self.container_spec = None
         self.pipeline_spec = None
 
+        def validate_placeholder_types(
+                component_spec: structures.ComponentSpec) -> None:
+            inputs_dict = component_spec.inputs or {}
+            outputs_dict = component_spec.outputs or {}
+            for arg in itertools.chain(
+                (component_spec.implementation.container.command or []),
+                (component_spec.implementation.container.args or [])):
+                check_primitive_placeholder_is_used_for_correct_io_type(
+                    inputs_dict, outputs_dict, arg)
+
         if component_spec.implementation.container is not None:
+            validate_placeholder_types(component_spec)
             self.container_spec = self._extract_container_spec_and_convert_placeholders(
                 component_spec=component_spec)
         elif component_spec.implementation.importer is not None:
@@ -452,3 +464,56 @@ class PipelineTask:
                 )
             self._task_spec.dependent_tasks.append(task.name)
         return self
+
+
+# TODO: this function should ideally be in the function kfp.components.structures.check_placeholder_references_valid_io_name, which does something similar, but this causes the exception to be raised at component definition time, rather than compile time. This would break tests that load v1 component YAML, even though that YAML is invalid.
+def check_primitive_placeholder_is_used_for_correct_io_type(
+    inputs_dict: Dict[str, structures.InputSpec],
+    outputs_dict: Dict[str, structures.OutputSpec],
+    arg: Union[placeholders.Placeholder, Any],
+):
+    """Validates input/output placeholders refer to an input/output with an
+    appropriate type for the placeholder. This should only apply to components
+    loaded from v1 component YAML, where the YAML is authored directly. For v2
+    YAML, this is encapsulated in the DSL logic which does not permit writing
+    incorrect placeholders.
+
+    Args:
+        inputs_dict: The existing input names.
+        outputs_dict: The existing output names.
+        arg: The command line element, which may be a placeholder.
+    """
+
+    if isinstance(arg, placeholders.InputValuePlaceholder):
+        input_name = arg.input_name
+        if not type_utils.is_parameter_type(inputs_dict[input_name].type):
+            raise TypeError(
+                f'Input "{input_name}" with type '
+                f'"{inputs_dict[input_name].type}" cannot be paired with '
+                'InputValuePlaceholder.')
+
+    elif isinstance(
+            arg,
+        (placeholders.InputUriPlaceholder, placeholders.InputPathPlaceholder)):
+        input_name = arg.input_name
+        if type_utils.is_parameter_type(inputs_dict[input_name].type):
+            raise TypeError(
+                f'Input "{input_name}" with type '
+                f'"{inputs_dict[input_name].type}" cannot be paired with '
+                f'{arg.__class__.__name__}.')
+
+    elif isinstance(arg, placeholders.OutputUriPlaceholder):
+        output_name = arg.output_name
+        if type_utils.is_parameter_type(outputs_dict[output_name].type):
+            raise TypeError(
+                f'Output "{output_name}" with type '
+                f'"{outputs_dict[output_name].type}" cannot be paired with '
+                f'{arg.__class__.__name__}.')
+    elif isinstance(arg, placeholders.IfPresentPlaceholder):
+        for arg in itertools.chain(arg.then or [], arg.else_ or []):
+            check_primitive_placeholder_is_used_for_correct_io_type(
+                inputs_dict, outputs_dict, arg)
+    elif isinstance(arg, placeholders.ConcatPlaceholder):
+        for arg in arg.items:
+            check_primitive_placeholder_is_used_for_correct_io_type(
+                inputs_dict, outputs_dict, arg)
